fix(telescope): allow vim.notify to post in command line after closing Telescope prompt
Previously, when closing the Telescope prompt after copying a commit, the message from vim.notify wouldn't show up in the command line, but it would be visible in :messages. Adding a delay with vim.fn.timer_start fixes this issue, so the message from vim.notify does show up in the command line.