feat(on_attach): always get feedback for completion and formatting
Previously the keymap for completion and user command for formatting
wouldn't be defined if completion or formatting wasn't available,
respectively. But now they're defined anyway, so a feedback message will
be displayed if they're used but unavailable.

The corresponding autocommands are still only defined if completion or
formatting is available.